# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "auth/login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email parameter in path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "test@mail.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password parameter in path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1234"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns some sample user things",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "List of users",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "List of users",
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/api/v1/auth/check": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "auth/check",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "The token parameter in path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJsdW1lbi1qd3QiLCJzdWIiOiI3ZmNjZTIwNy0xOGE2LTRlOWUtYjUwNS02YTZkMmMwZjhjM2QiLCJpYXQiOjE2MjYzMzQ5NzksImV4cCI6MTYyNjMzODU3OX0.ry2RohXZqrmjV6HxI8MYKamBhpA2cWEuvgYfAW9hnIU"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns some sample user things",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "List of users",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "List of users",
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "auth/logout",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "The token parameter in path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJsdW1lbi1qd3QiLCJzdWIiOiI3ZmNjZTIwNy0xOGE2LTRlOWUtYjUwNS02YTZkMmMwZjhjM2QiLCJpYXQiOjE2MjYzMzQ5NzksImV4cCI6MTYyNjMzODU3OX0.ry2RohXZqrmjV6HxI8MYKamBhpA2cWEuvgYfAW9hnIU"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns some sample user things",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "List of users",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "List of users",
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "description": "List of users",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Id User",
                        "type": "string",
                        "example": "e9e341be-218a-4701-987c-6b5ff7f9b0fd"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name User",
                        "type": "string",
                        "example": "sanusi"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email User",
                        "type": "string",
                        "example": "test@mail.com"
                    },
                    "token": {
                        "description": "Token User",
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJsdW1lbi1qd3QiLCJzdWIiOiI3ZmNjZTIwNy0xOGE2LTRlOWUtYjUwNS02YTZkMmMwZjhjM2QiLCJpYXQiOjE2MjYzMzQ5NzksImV4cCI6MTYyNjMzODU3OX0.ry2RohXZqrmjV6HxI8MYKamBhpA2cWEuvgYfAW9hnIU"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        }
    }
}